EVChargingSimulator/requirements.txt
alabaster==0.7.12
asgiref==3.4.0
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coverage==5.5
Django==3.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
djangorestframework==3.12.4
docutils==0.17.1
idna==3.2
imagesize==1.2.0
install==1.3.4
Jinja2==3.0.1



mock==4.0.3
mysqlclient==2.0.3

packaging==21.0
Pygments==2.10.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
scipy==1.7.0

snowballstemmer==2.1.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
urllib3==1.26.6
wincertstore==0.2

gunicorn==20.1.0

django-environ==0.8.1